+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Date format: DD/MM/YYYY
 
+## 17/07/2022
+
+- Added the `weapon_to_fire_group` collection.
+- Fixed the `count` verb returning incorrect values for certain collections.
+- Fixed the `count` verb not redirecting to DBG Census for unknown collections.
+
 ## 16/07/2022
 
 - Added automatic redirect to DBG Census for unknown collections.